Edits to README.md
Changed "built in" to "built-in" in document's first header.
Edited first sentence of document ("Handling dynamic...") by making noun clause subject into sentence object. It's easier to read if a long clause is moved to the end of a sentence. It makes it easier to understand what should be easy and robust now.
Capitalized Redbird to be consistent with rest of document.